Updated GameInfo and Added Workshop
GamInfo was updated with naming fixes and removal of levelCompletion check in the Update function. Now WorkShop file is attached to all workshops, where each of them checks  the collection of Sparkles. When certain amount was added, level is considered completed and UnityEvent is send to the GameInfo where LevelCompletedFunction tracks the progress.